requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.36.1
botocore==1.36.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.7
fastjsonschema==2.21.1
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
load-dotenv==0.1.0
logger==1.4
logging==0.4.9.6
logging-config==1.0.4
mangum==0.19.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.1
ollama==0.4.5
openai==1.60.2
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.53
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.1
schedule==1.2.2
sentry-sdk==2.19.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
